Reorganized Program class and add book_tickets
Each method of Program class now represents a stage of User Interface in the application. Starting from the start() method, control will be passed onto subsequent methods depending on user input, ultimately terminating if any of the methods fail or if the user provides the appropriate command to end the program (trying to standardize as "3" throughout).

Also added book_tickets method to allow user to specify number of tickets to reserve and call screening method to handle the booking and searching for the appropriate number of seats.