Modified alphanumeric check segment
Now abstracted to its own function, which returns a tuple of Falsy strings if the alphanumeric string provided is not suitable in this program's context.

Also moved message generation when selecting seats to Program class from Screening class, to reconcile user-facing text being located in the Program class. A little less than ideal as the method currently has to access the Booking class but this may not actually be a problem.